refactor(src) conflate trait code
