+++ src/tad/WPBrowser/Traits/WithSeparateProcessChecks.php
@@ -11,12 +11,55 @@
 
 namespace tad\WPBrowser\Traits;
 
+use Codeception\Exception\TestRuntimeException;
 use function tad\WPBrowser\phpunitVersion;
 
 if ( version_compare(phpunitVersion(),'7.0.0','>=') ) {
-	// Load the version compatible with PHPUnit >= 7.0.0.
-	require_once __DIR__ . '/_WithSeparateProcessChecksPHPUnitGte7.php';
+	/**
+	 * Trait WithSeparateProcessChecks
+	 *
+	 * @since   TBD
+	 *
+	 * @package tad\WPBrowser\Traits
+	 */
+	trait WithSeparateProcessChecks
+	{
+		/**
+		 * Overrides the base test case implementation to check the separate process configuration for the test
+		 * case is correct.
+		 *
+		 * @param bool $runTestInSeparateProcess Whether the test should run in a separate process or not.
+		 *
+		 * @throws TestRuntimeException If the test method, or test case, is configured to run in a separate process
+		 *                              preserving the global state.
+		 */
+		public function setRunTestInSeparateProcess(bool $runTestInSeparateProcess): void
+		{
+			parent::setRunTestInSeparateProcess($runTestInSeparateProcess);
+			$this->checkSeparateProcessConfiguration();
+		}
+	}
 } else {
-	// Load the version compatible with PHPUnit < 7.0.0.
-	require_once __DIR__ . '/_WithSeparateProcessChecksPHPUnitLt7.php';
+	/**
+	 * Trait WithSeparateProcessChecks
+	 *
+	 * @since   TBD
+	 *
+	 * @package tad\WPBrowser\Traits
+	 */
+	trait WithSeparateProcessChecks {
+		/**
+		 * Overrides the base test case implementation to check the separate process configuration for the test
+		 * case is correct.
+		 *
+		 * @param bool $runTestInSeparateProcess Whether the test should run in a separate process or not.
+		 *
+		 * @throws TestRuntimeException If the test method, or test case, is configured to run in a separate process
+		 *                              preserving the global state.
+		 */
+		public function setRunTestInSeparateProcess( $runTestInSeparateProcess ) {
+			parent::setRunTestInSeparateProcess( $runTestInSeparateProcess );
+			$this->checkSeparateProcessConfiguration();
+		}
+	}
 }
